+++ paddleseg/utils/__init__.py
@@ -19,4 +19,4 @@ from .env import seg_env, get_sys_env
 from .utils import *
 from .timer import TimeAverager, calculate_eta
 from . import visualize
-from .ema import EMA
+from .ema import *
